+++ backend/src/test_helpers.ts
@@ -1,0 +1,86 @@
+import { HlcTimestamp } from './hybrid_logical_clock';
+import { Store } from './store';
+import { Voter, VoterCheckInEvent, EventType, PollbookEvent } from './types';
+
+export function createVoter(
+  voterId: string,
+  firstName: string,
+  lastName: string
+): Voter {
+  return {
+    voterId,
+    firstName,
+    lastName,
+    middleName: '',
+    suffix: '',
+    streetNumber: '',
+    addressSuffix: '',
+    houseFractionNumber: '',
+    streetName: '',
+    state: '',
+    apartmentUnitNumber: '',
+    addressLine2: '',
+    addressLine3: '',
+    postalCityTown: '',
+    postalZip5: '',
+    zip4: '',
+    mailingStreetNumber: '',
+    mailingSuffix: '',
+    mailingHouseFractionNumber: '',
+    mailingStreetName: '',
+    mailingApartmentUnitNumber: '',
+    mailingAddressLine2: '',
+    mailingAddressLine3: '',
+    mailingCityTown: '',
+    mailingState: '',
+    mailingZip5: '',
+    mailingZip4: '',
+    party: '',
+    district: '',
+  };
+}
+export function createVoterCheckInEvent(
+  localEventId: number,
+  machineId: string,
+  voterId: string,
+  hlcTimestamp: HlcTimestamp
+): VoterCheckInEvent {
+  const timestamp = new Date().toISOString();
+  return {
+    localEventId,
+    type: EventType.VoterCheckIn,
+    machineId,
+    timestamp: hlcTimestamp,
+    voterId,
+    checkInData: {
+      timestamp,
+      identificationMethod: {
+        type: 'photoId',
+        state: 'nh',
+      },
+      machineId,
+    },
+  };
+}
+export function syncEventsFromTo(from: Store, to: Store): PollbookEvent[] {
+  let keepSyncing = true;
+  const allEvents: PollbookEvent[] = [];
+  while (keepSyncing) {
+    const lastSyncHeads = to.getLastEventSyncedPerNode();
+    const { events, hasMore } = from.getNewEvents(lastSyncHeads);
+    to.saveRemoteEvents(events);
+    allEvents.push(...events);
+    keepSyncing = hasMore;
+  }
+  return allEvents;
+}
+
+export function syncEventsForAllPollbooks(pollbooks: Store[]): void {
+  for (const from of pollbooks) {
+    for (const to of pollbooks) {
+      if (from !== to) {
+        syncEventsFromTo(from, to);
+      }
+    }
+  }
+}
